Collision work and scene changing view fix
Added pointer for the frame velocity, in the Collider and made the
Entity automatically fill it.
Added more content to both sides of the collision code, the Area's loop
and Colliders collision.
Added more layering to the scene loading pipeline, allowing a universal
scene section, before scene individual stuff.
Fixed the view bug, when changing from Area to Mainmenu, it wasn't
resetting the view, ending up in faulty placements.